requirements.txt
asgiref==3.6.0
boto3==1.26.129
botocore==1.29.129
coverage==7.2.5
dj-database-url==0.5.0
Django==3.2
django-active-link==0.1.8
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
django-daterangefilter==1.0.0
django-phonenumber-field==7.1.0
django-shortcuts==1.6
django-storages==1.13.2
django-summernote==0.8.20.0
gunicorn==20.1.0
jmespath==1.0.1
oauthlib==3.2.2
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2==2.9.6
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
s3transfer==0.6.1
sqlparse==0.4.4
stripe==5.4.0
